+++ CharacterOptionsPlus/CharacterOptionsPlus/Spells/FreezingSphere.cs
@@ -1,9 +1,11 @@
 ﻿using BlueprintCore.Actions.Builder;
 using BlueprintCore.Actions.Builder.ContextEx;
+using BlueprintCore.Blueprints.Configurators;
 using BlueprintCore.Blueprints.CustomConfigurators.UnitLogic.Abilities;
 using BlueprintCore.Blueprints.References;
 using BlueprintCore.Conditions.Builder;
 using BlueprintCore.Conditions.Builder.ContextEx;
+using BlueprintCore.Utils.Assets;
 using BlueprintCore.Utils.Types;
 using CharacterOptionsPlus.Util;
 using Kingmaker.Blueprints.Classes.Spells;
@@ -16,6 +18,7 @@ using Kingmaker.UnitLogic.Abilities.Components;
 using Kingmaker.Utility;
 using System;
 using System.Collections.Generic;
+using UnityEngine;
 using static Kingmaker.UnitLogic.Commands.Base.UnitCommand;
 using static Kingmaker.Visual.Animation.Kingmaker.Actions.UnitAnimationActionCastSpell;
 using static TabletopTweaks.Core.MechanicsChanges.MetamagicExtention;
@@ -30,8 +33,11 @@ namespace CharacterOptionsPlus.Spells
     internal const string DisplayName = "FreezingSphere.Name";
     private const string Description = "FreezingSphere.Description";
 
-    private const string IconPrefix = "assets/icons/";
-    private const string IconName = IconPrefix + "burningdisarm.png";
+    private const string ProjectileName = "FreezingSphere.Projectile";
+
+    // Effect from Kinetic_IceSphere00
+    private const string AreaEffectFxSource = "b23302f818cab9a4f9b57b821195ed01";
+    private const string AreaEffectFx = "8bb1eeb9-1255-4c36-bf22-7c01a7af4ad7";
 
     private static readonly ModLogger Logger = Logging.GetLogger(FeatureName);
 
@@ -54,12 +60,20 @@ namespace CharacterOptionsPlus.Spells
     {
       Logger.Log($"Configuring {FeatureName} (disabled)");
 
+      ProjectileConfigurator.New(ProjectileName, Guids.FreezingSphereProjectile).Configure();
       AbilityConfigurator.New(FeatureName, Guids.FreezingSphereSpell).Configure();
     }
 
     private static void ConfigureEnabled()
     {
       Logger.Log($"Configuring {FeatureName}");
+
+      AssetTool.RegisterDynamicPrefabLink(AreaEffectFx, AreaEffectFxSource, ModifyFx);
+      var projectile = ProjectileConfigurator.New(ProjectileName, Guids.FreezingSphereProjectile)
+        .CopyFrom(ProjectileRefs.Kinetic_IceSphere00_Projectile)
+        .ModifyProjectileHit(hit => hit.HitFx = new() { AssetId = AreaEffectFx})
+        .SetSpeed(10f)
+        .Configure();
 
       var availableMetamagic =
         new List<Metamagic>()
@@ -109,7 +123,7 @@ namespace CharacterOptionsPlus.Spells
           SpellDescriptor.Cold)
         .SetDisplayName(DisplayName)
         .SetDescription(Description)
-        .SetIcon(IconName)
+        .SetIcon(AbilityRefs.BlizzardBlastAbility.Reference.Get().Icon)
         .SetRange(AbilityRange.Long)
         .AllowTargeting(enemies: true, point: true)
         .SetAnimation(CastAnimationStyle.Point)
@@ -122,12 +136,24 @@ namespace CharacterOptionsPlus.Spells
         .AddAbilityDeliverProjectile(
           type: AbilityProjectileType.Simple,
           projectiles: new() { ProjectileRefs.Kinetic_IceSphere00_Projectile.ToString() })
-        .AddAbilityEffectRunAction(
-          alwaysSelective
-            ? ActionsBuilder.New().Conditional(ConditionsBuilder.New().IsEnemy(), ifTrue: attack)
-            : attack)
+        .AddAbilityTargetsAround(
+          radius: 40.Feet(),
+          targetType:  alwaysSelective ? TargetType.Enemy : TargetType.Any,
+          spreadSpeed: 45.Feet())
+        .AddAbilityEffectRunAction(attack)
         .Configure();
     }
 
+    private static void ModifyFx(GameObject iceSphere)
+    {
+      iceSphere.transform.localScale = new(1.75f, 1.0f, 1.75f);
+      iceSphere.transform.Find("StonesBig").localScale = new(1.5f, 1.0f, 1.5f);
+      iceSphere.transform.Find("IceSpike").localScale = new(3.5f, 1.0f, 3.5f);
+      iceSphere.transform.Find("Shockwave_FastBottom").localScale = new(3.5f, 1.0f, 3.5f);
+      //iceSphere.transform.Find("ProjectorDecal (1)_Start").localScale = new(1.5f, 1.0f, 1.5f);
+      iceSphere.transform.Find("Splashes (1)").localScale = new(1.75f, 1.0f, 1.75f);
+      iceSphere.transform.Find("DropsSmall").localScale = new(1.75f, 1.0f, 1.75f);
+      iceSphere.transform.Find("Rescale").localScale = new(1.75f, 1.0f, 1.75f);
+    }
   }
 }